Refactor store structure
