+++ src/components/footer/index.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 
 import { i18n, locales } from '@i18n';
-import { AuthActionType } from '@store/auth';
+import { AuthActionType } from '@store/enums';
 
 import './index.scss';
