--- src/store/auth/index.ts
@@ -1,5 +1,0 @@
-export * from './enums';
-export * from './interfaces';
-export * from './reducer';
-export { default as authReducer } from './reducer';
-export * from './sagas';
