+++ src/store/interfaces.ts
@@ -1,0 +1,5 @@
+import { store } from '..';
+
+export * from './branches/auth/interfaces';
+
+export type RootState = ReturnType<typeof store.getState>;
