+++ src/store/root-reducer.ts
@@ -2,10 +2,10 @@ import { History } from 'history';
 import { connectRouter } from 'connected-react-router';
 import { Reducer, combineReducers } from 'redux';
 
-import { authReducer } from './auth';
+import auth from '@store/branches/auth/reducer';
 
 export default (history: History<any>): Reducer =>
 	combineReducers({
-		auth: authReducer,
+		auth,
 		router: connectRouter(history)
 	});
