+++ src/store/sagas.ts
@@ -1,3 +1,3 @@
-import { loginSaga, logoutSaga, localeSaga, passwordResetSaga } from './auth';
+import { loginSaga, logoutSaga, localeSaga, passwordResetSaga } from '@store/branches/auth/sagas';
 
 export default [loginSaga, logoutSaga, localeSaga, passwordResetSaga];
