+++ src/utilities/hooks.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
 
-import { RootState } from '@store/index';
+import { RootState } from '@store/interfaces';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
